+++ client/src/components/Dashboard/Achievements.js
@@ -1,0 +1,9 @@
+const Achievements = () => {
+    return (
+        <section>
+            {/* Achievements section */}
+        </section>
+    )
+}
+
+export default Achievements;
