+++ client/src/components/Dashboard/Progess.js
@@ -1,0 +1,9 @@
+const Progress = () => {
+    return (
+        <section>
+            {/* graph.js implementation */}
+        </section>
+    )
+}
+
+export default Progress;
