+++ client/src/components/Dashboard/UserInfo.js
@@ -1,0 +1,9 @@
+const UserInfo = () => {
+    return (
+        <section>
+            {/* User Info Section. IE: Name, Date of account created(?) tests complete(?) etc */}
+        </section>
+    )
+}
+
+export default UserInfo;
